HBX-1694: Move class 'org.hibernate.tool.hbm2x.TemplateHelper' to 'org.hibernate.tool.internal.export.common.TemplateHelper'
